telegram_bot/requirements.txt
python-telegram-bot==13.4
requests==2.25.1
